revamped bot.py
