+++ cogs/bot.py
@@ -3,27 +3,20 @@ from discord.ext import commands
 import os
 import sys
 import time
-import math
 
-from helper import create_embed, get_settings, save_settings, get_channel, get_role, is_guild_owner, format_time
-from constants import VERSION_LOGS, VC_LANGUAGES, VC_ACCENTS
-
-def get_first_n_items(dictionary, number):
-    new_dictionary = {}
-    for index in list(dictionary)[:number]:
-        new_dictionary[index] = dictionary.get(index)
-    return new_dictionary
+from helper import create_embed, get_guild_data, save_guild_data, get_object, is_guild_owner, format_time, get_first_n_items
+from constants import VERSION_LOGS, VC_LANGUAGES, VC_ACCENTS, MAX_LEADERBOARD_FIELDS
 
 class bot(commands.Cog, description = "Bot management and settings."):
     def __init__(self, client):
         self.client = client
         self.uptime = time.time()
 
-    @commands.command(aliases = ["run"], description = "Runs code through the bot.", brief = "bot creator or server owner")
+    @commands.command(description = "Runs code through the bot.", brief = "server owner")
     @commands.check_any(commands.is_owner(), is_guild_owner())
-    async def execute(self, context, *, code):
+    async def run(self, context, *, code):
         response = await context.send(embed = create_embed({
-            "title": "Executing code...",
+            "title": "Running code...",
             "color": discord.Color.gold(),
         }, {
             "Code": code,
@@ -31,19 +24,18 @@ class bot(commands.Cog, description = "Bot management and settings."):
 
         try:
             exec(code)
+            await response.edit(embed = create_embed({
+                "title": "Ran code",
+                "color": discord.Color.green(),
+            }, {
+                "Code": code,
+            }))
         except Exception as error_message:
             await response.edit(embed = create_embed({
-                "title": "Could not execute code",
+                "title": "Could not run code",
                 "color": discord.Color.red(),
             }, {
                 "Error Message": error_message,
-                "Code": code,
-            }))
-        else:
-            await response.edit(embed = create_embed({
-                "title": "Code executed",
-                "color": discord.Color.green(),
-            }, {
                 "Code": code,
             }))
 
@@ -57,6 +49,10 @@ class bot(commands.Cog, description = "Bot management and settings."):
 
         try:
             os.system("cls" if os.name == "nt" else "clear")
+            await response.edit(embed = create_embed({
+                "title": "Terminal cleared",
+                "color": discord.Color.green(),
+            }))
         except Exception as error_message:
             await response.edit(embed = create_embed({
                 "title": "Could not clear terminal",
@@ -64,47 +60,32 @@ class bot(commands.Cog, description = "Bot management and settings."):
             }, {
                 "Error Message": error_message,
             }))
-        else:
-            await response.edit(embed = create_embed({
-                "title": "Terminal cleared",
-                "color": discord.Color.green(),
-            }))
-
-    @commands.command(description = "Changes the bot's activity in its profile.", brief = "bot creator or server owner")
+    
+    @commands.command(description = "Changes the bot's activity in its profile.", brief = "server owner")
     @commands.check_any(commands.is_owner(), is_guild_owner())
-    async def changeactivity(self, context, *, activity: str = ""):
-        activity = activity.lower()
-
+    async def changeactivity(self, context, *, activity = None):
         response = await context.send(embed = create_embed({
-            "title": "Changing the bot's activity",
+            "title": f"Changing the bot's activity to {activity}",
             "color": discord.Color.gold(),
-        }, {
-            "Activity": activity or "None",
         }))
 
         try:
-            await self.client.change_presence(activity = discord.Game(name = activity))
+            await self.client.change_presence(activity = discord.Game(name = activity or ""))
+            await response.edit(embed = create_embed({
+                "title": f"Changed bot's activity to {activity}",
+                "color": discord.Color.green(),
+            }))
         except Exception as error_message:
             await response.edit(embed = create_embed({
-                "title": "Could not change the bot's activity",
+                "title": f"Could not change the bot's activity to {activity}",
                 "color": discord.Color.red(),
             }, {
                 "Error Message": error_message,
-                "Activity": activity or "None",
-            }))
-        else:
-            await response.edit(embed = create_embed({
-                "title": "Changed bot's activity",
-                "color": discord.Color.green(),
-            }, {
-                "Activity": activity or "None",
             }))
 
-    @commands.command(description = "Changes the bot's status in its profile. Possible statuses are online, offline, dnd, and idle.", brief = "bot creator or server owner")
+    @commands.command(description = "Changes the bot's status in its profile. Possible statuses are online, offline, dnd, and idle.", brief = "server owner")
     @commands.check_any(commands.is_owner(), is_guild_owner())
     async def changestatus(self, context, *, status: str = "online"):
-        status = status.lower()
-
         response = await context.send(embed = create_embed({
             "title": f"Changing the bot's status to {status}",
             "color": discord.Color.gold(),
@@ -112,6 +93,10 @@ class bot(commands.Cog, description = "Bot management and settings."):
 
         try:
             await self.client.change_presence(status = discord.Status[status])
+            await response.edit(embed = create_embed({
+                "title": f"Changed bot's status to {status}",
+                "color": discord.Color.green(),
+            }))
         except Exception as error_message:
             await response.edit(embed = create_embed({
                 "title": f"Could not change the bot's status to {status}",
@@ -119,477 +104,16 @@ class bot(commands.Cog, description = "Bot management and settings."):
             }, {
                 "Error Message": error_message,
             }))
-        else:
-            await response.edit(embed = create_embed({
-                "title": f"Changed bot's status to {status}",
-                "color": discord.Color.green(),
-            }))
 
-    @commands.command(description = "Stops and starts the bot if it is being hosted on a server. Stops the bot if it is being hosted locally.", brief = "bot creator or server owner")
+    @commands.command(description = "Restarts the bot if it is being hosted on a server. Stops the bot if it is being hosted locally.", brief = "server owner")
     @commands.check_any(commands.is_owner(), is_guild_owner())
     async def restart(self, context):
         await context.send(embed = create_embed({
             "title": "Restarting bot, no confirmation message after completion...",
             "color": discord.Color.gold(),
         }))
+
         sys.exit()
-
-    @commands.command(aliases = ["set"], description = "Changes a server specific setting in the data stores.", brief = "specific")
-    async def setsettings(self, context, name: str, *, value = None):
-        response = await context.send(embed = create_embed({
-            "title": "Changing settings...",
-            "color": discord.Color.gold(),
-        }, {
-            "Name": name,
-            "Value": value,
-        }))
-
-        try:
-            settings = get_settings(context.guild.id)
-            name = name.lower()
-            if name == "join_channel":
-                if not context.author.guild_permissions.administrator and not await self.client.is_owner(context.author):
-                    await response.edit(embed = create_embed({
-                        "title": f"You don't have administrator or bot creator permissions",
-                        "color": discord.Color.red(),
-                    }))
-                    return
-
-                if not value or value.lower() == "none":
-                    settings["join_channel"] = None
-                    save_settings(settings)
-
-                    await response.edit(embed = create_embed({
-                        "title": "Removed join channel",
-                        "color": discord.Color.green(),
-                    }))
-                else:
-                    channel = get_channel(context.guild.text_channels, value)
-                    if channel:
-                        settings["join_channel"] = channel.id
-                        save_settings(settings)
-
-                        await response.edit(embed = create_embed({
-                            "title": f"Set join channel to {channel}",
-                            "color": discord.Color.green(),
-                        }))
-                    else:
-                        await response.edit(embed = create_embed({
-                            "title": f"{value} is not a valid channel",
-                            "color": discord.Color.red(),
-                        }))
-            elif name == "default_role":
-                if not context.author.guild_permissions.administrator and not await self.client.is_owner(context.author):
-                    await response.edit(embed = create_embed({
-                        "title": f"You don't have administrator or bot creator permissions",
-                        "color": discord.Color.red(),
-                    }))
-                    return
-
-                if not value or value.lower() == "none":
-                    settings["default_role"] = None
-                    save_settings(settings)
-
-                    await response.edit(embed = create_embed({
-                        "title": "Removed default role",
-                        "color": discord.Color.green(),
-                    }))
-                else:
-                    role = get_role(context.guild.roles, value)
-                    if role:
-                        settings["default_role"] = role.id
-                        save_settings(settings)
-
-                        await response.edit(embed = create_embed({
-                            "title": f"Set default role to {role}",
-                            "color": discord.Color.green(),
-                        }))
-                    else:
-                        await response.edit(embed = create_embed({
-                            "title": f"{value} is not a valid role",
-                            "color": discord.Color.red(),
-                        }))
-            
-            elif name == "acas_channel":
-                if not context.author.guild_permissions.administrator and not await self.client.is_owner(context.author):
-                    await response.edit(embed = create_embed({
-                        "title": f"You don't have administrator or bot creator permissions",
-                        "color": discord.Color.red(),
-                    }))
-                    return
-
-                if not value or value.lower() == "none":
-                    settings["acas_channel"] = None
-                    save_settings(settings)
-
-                    await response.edit(embed = create_embed({
-                        "title": "Removed ACAS channel",
-                        "color": discord.Color.green(),
-                    }))
-                else:
-                    channel = get_channel(context.guild.text_channels, value)
-                    if channel:
-                        settings["acas_channel"] = channel.id
-                        save_settings(settings)
-
-                        await response.edit(embed = create_embed({
-                            "title": f"Set ACAS channel to {channel}",
-                            "color": discord.Color.green(),
-                        }))
-                    else:
-                        await response.edit(embed = create_embed({
-                            "title": f"{value} is not a valid channel",
-                            "color": discord.Color.red(),
-                        }))
-            elif name == "acas_role":
-                if not context.author.guild_permissions.administrator and not await self.client.is_owner(context.author):
-                    await response.edit(embed = create_embed({
-                        "title": f"You don't have administrator or bot creator permissions",
-                        "color": discord.Color.red(),
-                    }))
-                    return
-
-                if not value or value.lower() == "none":
-                    settings["acas_role"] = None
-                    save_settings(settings)
-
-                    await response.edit(embed = create_embed({
-                        "title": "Removed ACAS role",
-                        "color": discord.Color.green(),
-                    }))
-                else:
-                    role = get_role(context.guild.roles, value)
-                    if role:
-                        settings["acas_role"] = role.id
-                        save_settings(settings)
-
-                        await response.edit(embed = create_embed({
-                            "title": f"Set ACAS role to {role}",
-                            "color": discord.Color.green(),
-                        }))
-                    else:
-                        await response.edit(embed = create_embed({
-                            "title": f"{value} is not a valid role",
-                            "color": discord.Color.red(),
-                        })) 
-            elif name == "acas_enabled":
-                if not value or value.lower() == "false":
-                    settings["acas_enabled"] = False
-                    save_settings(settings)
-
-                    await response.edit(embed = create_embed({
-                        "title": "Disabled ACAS",
-                        "color": discord.Color.green(),
-                    }))
-                elif value.lower() == "true":
-                    settings["acas_enabled"] = True
-                    save_settings(settings)
-
-                    await response.edit(embed = create_embed({
-                        "title": "Enabled ACAS",
-                        "color": discord.Color.green(),
-                    }))
-                else:
-                    await response.edit(embed = create_embed({
-                        "title": f"{value} is not a valid boolean (true/false)",
-                        "color": discord.Color.red()
-                    }))
-                    return
-            
-            elif name == "prefix":
-                value = str(value)
-                if value:
-                    settings["prefix"] = value
-                    save_settings(settings)
-
-                    await response.edit(embed = create_embed({
-                        "title": f"Changed prefix to {value}",
-                        "color": discord.Color.green(),
-                    }))
-                else:
-                    await response.edit(embed = create_embed({
-                        "title": f"{value} could not be converted to a string",
-                        "color": discord.Color.red()
-                    }))
-                    return
-            
-            elif name == "vc_accent":
-                value = str(value)
-                if value:
-                    if not VC_ACCENTS.get(value):
-                        await response.edit(embed = create_embed({
-                            "title": f"{value} is not a valid accent",
-                            "color": discord.Color.red()
-                        }))
-                        return
-
-                    settings["vc_accent"] = value
-                    save_settings(settings)
-
-                    await response.edit(embed = create_embed({
-                        "title": f"Changed the bot's accent to {value}",
-                        "color": discord.Color.green(),
-                    }))
-                else:
-                    await response.edit(embed = create_embed({
-                        "title": f"{value} could not be converted to a string",
-                        "color": discord.Color.red()
-                    }))
-            elif name == "vc_language":
-                value = str(value)
-                if value:
-                    if not VC_LANGUAGES.get(value):
-                        await response.edit(embed = create_embed({
-                            "title": f"{value} is not a valid language",
-                            "color": discord.Color.red()
-                        }))
-                        return
-
-                    settings["vc_language"] = value
-                    save_settings(settings)
-
-                    await response.edit(embed = create_embed({
-                        "title": f"Changed the bot's language to {value}",
-                        "color": discord.Color.green(),
-                    }))
-                else:
-                    await response.edit(embed = create_embed({
-                        "title": f"{value} could not be converted to a string",
-                        "color": discord.Color.red()
-                    }))
-                    return
-            elif name == "vc_slow_mode":
-                if not value or value.lower() == "false":
-                    settings["vc_slow_mode"] = False
-                    save_settings(settings)
-
-                    await response.edit(embed = create_embed({
-                        "title": "Disabled bot slow mode",
-                        "color": discord.Color.green(),
-                    }))
-                elif value.lower() == "true":
-                    settings["vc_slow_mode"] = True
-                    save_settings(settings)
-
-                    await response.edit(embed = create_embed({
-                        "title": "Enabled bot slow mode",
-                        "color": discord.Color.green(),
-                    }))
-                else:
-                    await response.edit(embed = create_embed({
-                        "title": f"{value} is not a valid boolean (true/false)",
-                        "color": discord.Color.red()
-                    }))
-                    return
-            
-            elif name == "voice_exp":
-                if not context.author.guild_permissions.administrator and not await self.client.is_owner(context.author):
-                    await response.edit(embed = create_embed({
-                        "title": f"You don't have administrator or bot creator permissions",
-                        "color": discord.Color.red(),
-                    }))
-                    return
-
-                value = int(value)
-                if value:
-                    settings["voice_exp"] = value
-                    save_settings(settings)
-
-                    await response.edit(embed = create_embed({
-                        "title": f"Set voice EXP to {value}",
-                        "color": discord.Color.green(),
-                    }))
-                else:
-                    await response.edit(embed = create_embed({
-                        "title": f"{value} is not a valid integer",
-                        "color": discord.Color.red()
-                    }))
-                    return
-            elif name == "message_exp":
-                if not context.author.guild_permissions.administrator and not await self.client.is_owner(context.author):
-                    await response.edit(embed = create_embed({
-                        "title": f"You don't have administrator or bot creator permissions",
-                        "color": discord.Color.red(),
-                    }))
-                    return
-
-                value = int(value)
-                if value:
-                    settings["message_exp"] = value
-                    save_settings(settings)
-
-                    await response.edit(embed = create_embed({
-                        "title": f"Set message EXP to {value}",
-                        "color": discord.Color.green(),
-                    }))
-                else:
-                    await response.edit(embed = create_embed({
-                        "title": f"{value} is not a valid integer",
-                        "color": discord.Color.red()
-                    }))
-                    return
-            elif name == "level_dificulty":
-                if not context.author.guild_permissions.administrator and not await self.client.is_owner(context.author):
-                    await response.edit(embed = create_embed({
-                        "title": f"You don't have administrator or bot creator permissions",
-                        "color": discord.Color.red(),
-                    }))
-                    return
-            
-                value = int(value)
-                if value:
-                    settings["level_dificulty"] = value
-                    save_settings(settings)
-
-                    await response.edit(embed = create_embed({
-                        "title": f"Set level dificulty to {value}",
-                        "color": discord.Color.green(),
-                    }))
-                else:
-                    await response.edit(embed = create_embed({
-                        "title": f"{value} is not a valid integer",
-                        "color": discord.Color.red()
-                    }))
-                    return
-            elif name == "message_cooldown":
-                if not context.author.guild_permissions.administrator and not await self.client.is_owner(context.author):
-                    await response.edit(embed = create_embed({
-                        "title": f"You don't have administrator or bot creator permissions",
-                        "color": discord.Color.red(),
-                    }))
-                    return
-
-                value = int(value)
-                if value:
-                    settings["message_cooldown"] = value
-                    save_settings(settings)
-
-                    await response.edit(embed = create_embed({
-                        "title": f"Set message cooldown to {value}",
-                        "color": discord.Color.green(),
-                    }))
-                else:
-                    await response.edit(embed = create_embed({
-                        "title": f"{value} is not a valid integer",
-                        "color": discord.Color.red()
-                    }))
-                    return
-            elif name == "money_per_level":
-                if not context.author.guild_permissions.administrator and not await self.client.is_owner(context.author):
-                    await response.edit(embed = create_embed({
-                        "title": f"You don't have administrator or bot creator permissions",
-                        "color": discord.Color.red(),
-                    }))
-                    return
-
-                value = int(value)
-                if value:
-                    settings["money_per_level"] = value
-                    save_settings(settings)
-
-                    await response.edit(embed = create_embed({
-                        "title": f"Set money per level reward to ${value}",
-                        "color": discord.Color.green(),
-                    }))
-                else:
-                    await response.edit(embed = create_embed({
-                        "title": f"{value} is not a valid integer",
-                        "color": discord.Color.red()
-                    }))
-                    return
-            
-            else:
-                await response.edit(embed = create_embed({
-                    "title": f"{name} is not a valid setting",
-                    "color": discord.Color.red(),
-                }))
-        except Exception as error_message:
-            await response.edit(embed = create_embed({
-                "title": "Unable to change settings",
-                "color": discord.Color.red(),
-            }, {
-                "Error Message": error_message,
-                "Name": name,
-                "Value": value,
-            }))            
-
-    @commands.command(aliases = ["settings"], description = "Retrieves a list of server specific settings in the data store.")
-    async def getsettings(self, context):
-        response = await context.send(embed = create_embed({
-            "title": "Loading settings...",
-            "color": discord.Color.gold(),
-        }))
-
-        try:
-            settings = get_settings(context.guild.id)
-            await response.edit(embed = create_embed({
-                "title": "Settings",
-                "inline": True,
-            }, settings))
-        except Exception as error_message:
-            await response.edit(embed = create_embed({
-                "title": "Unable to load settings",
-                "color": discord.Color.red(),
-            }, {
-                "Error Message": error_message,
-            }))
-
-    @commands.command(description = "Gets specific data.")
-    async def get(self, context, name: str, *, value = None):
-        response = await context.send(embed = create_embed({
-            "title": f"Loading {name}...",
-            "color": discord.Color.gold()
-        }))
-
-        try:
-            if name == "vc_language":
-                if value:
-                    value = int(value)
-                else:
-                    value = 1
-
-                first_page = value * 25 - 25
-                last_page = value * 25
-
-                fields = {}
-                for key, language_name in enumerate(list(VC_LANGUAGES.keys())):
-                    if key >= first_page and key < last_page:
-                        fields[language_name] = VC_LANGUAGES[language_name]
-
-                await response.edit(embed = create_embed({
-                    "title": f"VC Languages (Page {value})",
-                    "inline": True,
-                }, fields))
-            elif name == "vc_accent":
-                if value:
-                    value = int(value)
-                else:
-                    value = 1
-
-                first_page = value * 25 - 25
-                last_page = value * 25
-
-                fields = {}
-                for key, language_name in enumerate(list(VC_ACCENTS.keys())):
-                    if key >= first_page and key < last_page:
-                        fields[language_name] = VC_ACCENTS[language_name]
-
-                await response.edit(embed = create_embed({
-                    "title": f"VC Accents (Page {value})",
-                    "inline": True,
-                }, fields))
-            else:
-                await response.edit(embed = create_embed({
-                    "title": f"{name} does not have any data",
-                    "color": discord.Color.red()
-                }))
-        except Exception as error_message:
-            await response.edit(embed = create_embed({
-                "title": f"Could not get {name}",
-                "color": discord.Color.red()
-            }, {
-                "Error Message": error_message
-            }))
 
     @commands.command(description = "Retrieves the Python and discord.py version the bot is running on.")
     async def version(self, context):
@@ -618,7 +142,7 @@ class bot(commands.Cog, description = "Bot management and settings."):
             }))
 
     @commands.command(description = "Retrieves the bot's uptime, connected servers, members watching and users watching.")
-    async def info(self, context):
+    async def botinfo(self, context):
         response = await context.send(embed = create_embed({
             "title": "Loading bot info...",
             "color": discord.Color.gold()
@@ -658,11 +182,32 @@ class bot(commands.Cog, description = "Bot management and settings."):
             }))
 
     @commands.command()
-    async def updatelog(self, context):
-        logs = get_first_n_items(VERSION_LOGS, 10)
-        await context.send(embed = create_embed({
-            "title": "Update Log",
-        }, logs))
+    async def updatelog(self, context, page: int = 1):
+        response = await context.send(embed = create_embed({
+            "title": f"Loading update log...",
+            "color": discord.Color.gold()
+        }))
+
+        try:
+            first_page = page * MAX_LEADERBOARD_FIELDS - MAX_LEADERBOARD_FIELDS
+            last_page = page * MAX_LEADERBOARD_FIELDS
+
+            fields = {}
+            for key, version in enumerate(list(VERSION_LOGS.keys())):
+                if key >= first_page and key < last_page:
+                    fields[version] = VERSION_LOGS[version]
+
+            await response.edit(embed = create_embed({
+                "title": "Update Log",
+                "footer": f"Page {page}"
+            }, fields))
+        except Exception as error_message:
+            create_embed({
+                "title": f"Could not load update log",
+                "color": discord.Color.red()
+            }, {
+                "Error Message": error_message
+            })
 
 def setup(client):
     client.add_cog(bot(client))
